front-end/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.2
django-cors-headers==4.3.1
pika==1.3.2
python-dotenv==0.16.0
psycopg2-binary==2.9.9
